Add files via upload
Developed a shopping application or e-commerce application which has login and public login features on the Python platform. The application that have been developed include categories, such as 3–4 for footwear, clothing, electronics, etc. It is possible to add and update categories in the application. Additionally, it allows you to add or remove items from your cart. Finally, the program supports a variety of payment options, including UPI and debit cards. This is a only backend implementation.